+++ packages/cli/__tests__/workflows/__system__/List/retrieveWorkflowDefinition/cli.workflows.list.retrieveWorkflowDefinition.system.test.ts
@@ -48,7 +48,7 @@ describe("Retrieve workflow definition cli system tests", () => {
     describe("List workflow definition file details", () => {
         beforeAll(async () => {
             // Upload files only for successful scenarios
-            await Upload.fileToUSSFile(REAL_SESSION, workflow, definitionFile, true);
+            await Upload.fileToUssFile(REAL_SESSION, workflow, definitionFile, { binary: true });
         });
         afterAll(async () => {
             let error;