+++ packages/cli/__tests__/workflows/__system__/workflow/cli.workflows.archive.workflow.system.test.ts
@@ -50,7 +50,7 @@ describe("Archive workflow cli system tests", () => {
     describe("Archive workflows", () => {
         beforeAll(async () => {
             // Upload files only for successful scenarios
-            await Upload.fileToUSSFile(REAL_SESSION, workflow, definitionFile, true);
+            await Upload.fileToUssFile(REAL_SESSION, workflow, definitionFile, { binary: true });
         });
         afterAll(async () => {
             let error;