+++ packages/workflows/__tests__/__system__/api/ListArchivedWorkflows.system.test.ts
@@ -72,7 +72,7 @@ describe("List archived workflows", () => {
     describe("Success Scenarios", () => {
         beforeAll(async () => {
             // Upload files only for successful scenarios
-            await Upload.fileToUSSFile(REAL_SESSION, workflow, definitionFile, true);
+            await Upload.fileToUssFile(REAL_SESSION, workflow, definitionFile, { binary: true });
         });
         afterAll(async () => {
             let error;