+++ packages/zostso/src/doc/zosmf/IZosmfMessages.ts
@@ -12,14 +12,6 @@
 /**
  * The z/OSMF Ping API error message parameters. See the z/OSMF REST API documentation for full details.
  * @export
- * @deprecated Use `IZosmfMessages` instead because it is spelled correctly
- * @interface IZosfmMessages
- */
-export type IZosfmMessages = IZosmfMessages;
-
-/**
- * The z/OSMF Ping API error message parameters. See the z/OSMF REST API documentation for full details.
- * @export
  * @interface IZosmfMessages
  */
 export interface IZosmfMessages {